refactor: types
